Module ext::chain_complex

Modules

chain_homotopy

finite_chain_complex

tensor_product_chain_complex

Structs

ChainHomotopy
A chain homotopy from $f to g$, or equivalently a null-homotopy of $h = f - g$. A chain map is a priori a collection of free module homomorphisms. However, instead of providing FreeModuleHomomorphism objects, the user is expected to give a function that computes the value of $h$ on each generator.

ChainMap
chain_maps is required to be non-empty

FiniteAugmentedChainComplex

FiniteChainComplex

StemIterator
An iterator returned by ChainComplex::iter_stem

TensorChainComplex

Enums

ChainComplexGrading

Traits

AugmentedChainComplex
An augmented chain complex is a map of chain complexes C -> D that is a quasi-isomorphism. We usually think of C as a resolution of D. The chain map must be a map of degree shift 0.

BoundedChainComplex
A bounded chain complex is a chain complex C for which C_s = 0 for all s >= max_s

ChainComplex
A chain complex is defined to start in degree 0. The min_degree is the min_degree of the modules in the chain complex, all of which must be the same.

CochainComplex

FreeChainComplex