Struct ext::chain_complex::ChainHomotopy

pub struct ChainHomotopy<'a, S: FreeChainComplex, T: ChainComplex, F: Fn(u32, i32, usize, &mut FpVector)> {
    source: &'a S,
    target: &'a T,
    shift_s: u32,
    shift_t: i32,
    map: F,
    lock: Mutex<()>,
    homotopies: OnceVec<FreeModuleHomomorphism<T::Module>>,
}

A chain homotopy from $f to g$, or equivalently a null-homotopy of $h = f - g$. A chain map is a priori a collection of free module homomorphisms. However, instead of providing FreeModuleHomomorphism objects, the user is expected to give a function that computes the value of $h$ on each generator.

Fields
source: &'a Starget: &'a Tshift_s: u32

The $s$ shift of the original chain map $f - g$.

shift_t: i32

The $t$ shift of the original chain map $f - g$.

map: F

A function that given (s, t, idx, result), adds (f - g)(x_{s, t, i}), to result.

lock: Mutex<()>homotopies: OnceVec<FreeModuleHomomorphism<T::Module>>

Homotopies, indexed by the filtration of the target of f - g.

Implementations

impl<'a, S: FreeChainComplex, T: ChainComplex<Algebra = S::Algebra>, F: Fn(u32, i32, usize, &mut FpVector)> ChainHomotopy<'a, S, T, F>

pub fn new(
source: &'a S,
target: &'a T,
shift_s: u32,
shift_t: i32,
map: F
) -> Self

pub fn prime(&self) -> ValidPrime

pub fn extend(&self, max_source_s: u32, max_source_t: i32)

Lift maps so that the chain homotopy is defined on (max_source_s, max_source_t).

pub fn extend_all(&self)

Lift maps so that the chain homotopy is defined on as many bidegrees as possible

fn extend_step(&self, source_s: u32, source_t: i32, scratch: &mut FpVector)

pub fn homotopy(&self, source_s: u32) -> &FreeModuleHomomorphism<T::Module>

pub fn into_homotopies(self) -> OnceVec<FreeModuleHomomorphism<T::Module>>

Into the vec of homotopies. This Vec is indexed by the homological degree of the target of f - g.

impl<'a, S: FreeChainComplex, T: ChainComplex<Algebra = S::Algebra> + Sync, F: Fn(u32, i32, usize, &mut FpVector) + Sync> ChainHomotopy<'a, S, T, F>

pub fn extend_all_concurrent(&self, bucket: &TokenBucket)