Module ext::module

Modules

block_structure

homomorphism

Structs

BCp

Dickson2

FDModule

FPModule

FreeModule
We have a linear enumeration of all generators among all degrees, and the index in this enumeration is the internal index.

FreeUnstableModule

HomModule

KFpn

ModuleFailedRelationError

OperationGeneratorPair

QuotientModule
Given a module module, this is the quotient of module by a collection of basis elements.

RealProjectiveSpace
This is $\mathbb{RP}_{\mathrm{min}}^{\mathrm{max}}$. The cohomology is the subquotient of $\mathbb{F}_2[x^\pm]$ given by elements of degree between min and max (inclusive)

SumModule

TensorModule

TruncatedModule
A module M where we quotient out everything above degree truncation

Enums

FiniteModule

Traits

BoundedModule

Module

PolynomialAlgebraModule

SteenrodModule

ZeroModule