Struct ext::module::homomorphism::BoundedModuleHomomorphism

pub struct BoundedModuleHomomorphism<S, T> where
    S: BoundedModule,
    T: Module<Algebra = <S as Module>::Algebra>,  {
    pub lock: Mutex<RawMutex, ()>,
    pub source: Arc<S>,
    pub target: Arc<T>,
    pub degree_shift: i32,
    pub matrices: BiVec<Matrix>,
    pub quasi_inverses: OnceBiVec<QuasiInverse>,
    pub kernels: OnceBiVec<Subspace>,
}

Fields
lock: Mutex<RawMutex, ()>source: Arc<S>target: Arc<T>degree_shift: i32matrices: BiVec<Matrix>quasi_inverses: OnceBiVec<QuasiInverse>kernels: OnceBiVec<Subspace>
Methods

impl<A, S, T> BoundedModuleHomomorphism<S, T> where
A: Algebra,
S: BoundedModule<Algebra = A>,
T: Module<Algebra = A>,

pub fn new(
source: Arc<S>,
target: Arc<T>,
degree_shift: i32
) -> BoundedModuleHomomorphism<S, T>

pub fn from_matrices(
source: Arc<S>,
target: Arc<T>,
degree_shift: i32,
matrices: BiVec<Matrix>
) -> BoundedModuleHomomorphism<S, T>

pub fn from<F>(f: &F) -> BoundedModuleHomomorphism<S, T> where
F: ModuleHomomorphism<Source = S, Target = T>,

pub fn replace_source<S_>(
self,
source: Arc<S_>
) -> BoundedModuleHomomorphism<S_, T> where
S_: BoundedModule<Algebra = A>,

This function replaces the source of the BoundedModuleHomomorphism and does nothing else. This is useful for changing the type of the source (but not the mathematical module itself). This is intended to be used in conjunction with BoundedModule::to_fd_module

pub fn replace_target<T_>(
self,
target: Arc<T_>
) -> BoundedModuleHomomorphism<S, T_> where
T_: BoundedModule<Algebra = A>,

See replace_source