Trait ext::module::homomorphism::ModuleHomomorphism

pub trait ModuleHomomorphism: 'static + Send + Sync where
    <Self::Target as Module>::Algebra == <Self::Source as Module>::Algebra, {
    type Source: Module;
    type Target: Module;

    const CUSTOM_QI: bool;

    fn source(&self) -> Arc<Self::Source>;

    fn target(&self) -> Arc<Self::Target>;

    fn degree_shift(&self) -> i32;

    fn apply_to_basis_element(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        input_degree: i32, 
        input_idx: usize
    );

    fn kernel(&self, degree: i32) -> &Subspace;

    fn quasi_inverse(&self, degree: i32) -> &QuasiInverse;

    fn compute_kernels_and_quasi_inverses_through_degree(&self, degree: i32);

    fn apply(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        input_degree: i32, 
        input: &FpVector
    ) { ... }

    fn prime(&self) -> ValidPrime { ... }

    fn min_degree(&self) -> i32 { ... }

    fn image(&self, degree: i32) -> &Option<Subspace> { ... }

    fn calculate_quasi_inverse(&self, degree: i32) -> QuasiInverse { ... }

    fn kernel_and_quasi_inverse(&self, degree: i32) -> (Subspace, QuasiInverse) { ... }

    fn get_matrix(&self, matrix: &mut Matrix, degree: i32) { ... }

    fn apply_quasi_inverse(
        &self, 
        result: &mut FpVector, 
        degree: i32, 
        input: &FpVector
    ) { ... }
}

Associated Types

type Source: Module

type Target: Module

Loading content...
Associated Constants

const CUSTOM_QI: bool

Loading content...
Required methods

fn source(&self) -> Arc<Self::Source>

fn target(&self) -> Arc<Self::Target>

fn degree_shift(&self) -> i32

fn apply_to_basis_element(
&self,
result: &mut FpVector,
coeff: u32,
input_degree: i32,
input_idx: usize
)

Calling this function when input_idx < source().dimension(input_degree) results in undefined behaviour. Implementations are encouraged to panic when this happens (this is usually the case because of out-of-bounds errors.

fn kernel(&self, degree: i32) -> &Subspace

fn quasi_inverse(&self, degree: i32) -> &QuasiInverse

fn compute_kernels_and_quasi_inverses_through_degree(&self, degree: i32)

Loading content...
Provided methods

fn apply(
&self,
result: &mut FpVector,
coeff: u32,
input_degree: i32,
input: &FpVector
)

fn prime(&self) -> ValidPrime

fn min_degree(&self) -> i32

fn image(&self, degree: i32) -> &Option<Subspace>

Returns the image of the module homomorphism in degree degree. If None, the image is the whole space.

fn calculate_quasi_inverse(&self, degree: i32) -> QuasiInverse

A version of kernel_and_quasi_inverse that, in fact, doesn't compute the kernel.

fn kernel_and_quasi_inverse(&self, degree: i32) -> (Subspace, QuasiInverse)

fn get_matrix(&self, matrix: &mut Matrix, degree: i32)

The (sliced) dimensions of matrix must be equal to source_dimension x target_dimension

fn apply_quasi_inverse(
&self,
result: &mut FpVector,
degree: i32,
input: &FpVector
)

Loading content...