ext
Module resolution

This module exports the Resolution object, which is a chain complex resolving a module. In particular, this contains the core logic that compute minimal resolutions.

Structs

MuResolution

A minimal resolution of a chain complex. The functions MuResolution::compute_through_stem and MuResolution::compute_through_bidegree extends the minimal resolution to the given bidegree.

SenderData 🔒

In MuResolution::compute_through_stem and MuResolution::compute_through_bidegree, we pass this struct around to inform the supervisor what bidegrees have been computed. We use an explicit struct instead of a tuple to avoid an infinite type problem.

Constants

MAX_NEW_GENS 🔒

This is the maximum number of new generators we expect in each bidegree. This affects how much space we allocate when we are extending our resolutions. Having more than this many new generators will result in a slowdown but not an error. It is relatively cheap to increment this number if needs be, but up to the 140th stem we only see at most 8 new generators.

Type Aliases

Resolution

UnstableResolution