Struct ext::resolution::Resolution

pub struct Resolution<CC: UnitChainComplex> {
    pub inner: Arc<ResolutionInner<CC>>,
    pub next_s: Mutex<u32>,
    pub next_t: Mutex<i32>,
    pub add_class: Option<AddClassFn>,
    pub add_structline: Option<AddStructlineFn>,
    filtration_one_products: Vec<(String, i32, usize)>,
    pub unit_resolution: Option<Weak<RwLock<Resolution<CC>>>>,
    pub unit_resolution_owner: Option<Arc<RwLock<Resolution<CC>>>>,
    product_names: HashSet<String>,
    product_list: Vec<Cocycle>,
    chain_maps_to_unit_resolution: OnceVec<OnceBiVec<OnceVec<ResolutionHomomorphismToUnit<CC>>>>,
    max_product_homological_degree: u32,
    pub self_maps: Vec<SelfMap<CC>>,
}

Fields

kernels - For each internal degree, store the kernel of the most recently calculated chain map as returned by generate_old_kernel_and_compute_new_kernel, to be used if we run resolve_through_degree again.

Fields
inner: Arc<ResolutionInner<CC>>next_s: Mutex<u32>next_t: Mutex<i32>add_class: Option<AddClassFn>add_structline: Option<AddStructlineFn>filtration_one_products: Vec<(String, i32, usize)>unit_resolution: Option<Weak<RwLock<Resolution<CC>>>>unit_resolution_owner: Option<Arc<RwLock<Resolution<CC>>>>product_names: HashSet<String>product_list: Vec<Cocycle>chain_maps_to_unit_resolution: OnceVec<OnceBiVec<OnceVec<ResolutionHomomorphismToUnit<CC>>>>max_product_homological_degree: u32self_maps: Vec<SelfMap<CC>>
Implementations

impl<CC: UnitChainComplex> Resolution<CC>

pub fn new(
complex: Arc<CC>,
add_class: Option<AddClassFn>,
add_structline: Option<AddStructlineFn>
) -> Self

pub fn new_with_inner(
inner: ResolutionInner<CC>,
add_class: Option<AddClassFn>,
add_structline: Option<AddStructlineFn>
) -> Self

pub fn resolve_through_bidegree(&self, max_s: u32, max_t: i32)

pub fn resolve_through_degree(&self, degree: i32)

fn step_after(&self, s: u32, t: i32)

fn compute_filtration_one_products(&self, target_s: u32, target_t: i32)

pub fn add_structline(
&self,
name: &str,
source_s: u32,
source_t: i32,
target_s: u32,
target_t: i32,
left: bool,
products: Vec<Vec<u32>>
)

pub fn max_computed_degree(&self) -> i32

pub fn max_computed_homological_degree(&self) -> u32

pub fn graded_dimension_vec(&self) -> Vec<Vec<usize>>

pub fn graded_dimension_string(&self) -> String

impl<CC: UnitChainComplex> Resolution<CC>

pub fn catch_up_products(&self)

This function computes the products between the element most recently added to product_list and the parts of Ext that have already been computed. This function should be called right after add_product, unless resolve_through_degree/resolve_through_bidegree has never been called.

This is made separate from add_product because extend_maps_to_unit needs a borrow of self, but add_product takes in a mutable borrow.

pub fn add_product(
&mut self,
s: u32,
t: i32,
class: Vec<u32>,
name: &str
) -> bool

The return value is whether the product was actually added. If the product is already present, we do nothing.

pub fn construct_unit_resolution(&mut self)

pub fn set_unit_resolution(&mut self, unit_res: Weak<RwLock<Resolution<CC>>>)

fn compute_products(&self, s: u32, t: i32, products: &[Cocycle])

Compute products whose result lie in degrees up to (s, t)

fn compute_product_step(&self, elt: &Cocycle, target_s: u32, target_t: i32)

Target = result of the product Source = multiplicand

fn construct_maps_to_unit(&self, s: u32, t: i32)

fn extend_maps_to_unit(&self, s: u32, t: i32)

This ensures the chain_maps_to_unit_resolution are defined such that we can compute products up to bidegree (s, t)

impl<CC: UnitChainComplex> Resolution<CC>

pub fn add_self_map(
&mut self,
s: u32,
t: i32,
name: &str,
map_data: Matrix
) -> bool

The return value is whether the self map was actually added. If the self map is already present, we do nothing.

fn compute_self_maps(&self, target_s: u32, target_t: i32)

We compute the products by self maps where the result has degree (s, t).

impl<CC: UnitChainComplex> Resolution<CC>

pub fn algebra(&self) -> Arc<<CC::Module as Module>::Algebra>

pub fn prime(&self) -> ValidPrime

pub fn module(
&self,
homological_degree: u32
) -> Arc<FreeModule<<CC::Module as Module>::Algebra>>

pub fn min_degree(&self) -> i32

pub fn differential(
&self,
s: u32
) -> Arc<FreeModuleHomomorphism<FreeModule<<CC::Module as Module>::Algebra>>>