ext

Module save

Structs

AbsolutePath

ChecksumReader

ChecksumWriter

In addition to checking the checksum, we also keep track of which files are open, and we delete the open files if the program is terminated halfway.

SaveFile

Enums

SaveDirectory

SaveKind

Functions

open_file 🔒

Open the file pointed to by path as a Box<dyn Read>. If the file does not exist, look for compressed versions.

open_files 🔒

A DashSet<PathBuf> of files that are currently opened and being written to. When calling this function for the first time, we set the ctrlc handler to delete currently opened files then exit.