Module ext::secondary

Macros

sub

unsub

Structs

MilnorClass

An element in the Milnor algebra

Constants

AY_CACHE

TWO

Functions

a_sigma_y

Compute the Y terms of $A(a, σ(b)σ(c))$

a_y

Computes $A(a, Y_{k, l})$ using a thread_local cache. This dispatches to a_y_cached that acts on individual Sq(R) instead of a list of them.

a_y_cached

Compute $A(Sq(R), Y_{k, l})$ where $a = Sq(R)$. This queries the cache and computes it using a_y_inner if not available.

a_y_inner

Actually computes $A(a, Y_{k, l})$ and returns the result.

can_compute

Whether picking δ₂ = 0 gives a valid secondary refinement. This requires

compute_a_dd

Computes $A(a, ddg)$

compute_c

Computes $C(g_i) = A(c_i^j, dd g_j)$.

compute_delta

A non-concurrent version for computing delta. In practice the concurrent version will be used, and this function should have clear logic rather than being optimal.

compute_delta_concurrent

read_saved_data

Type Definitions

FMH

Resolution