Module ext::utils

Structs

Config
A config object is an object that specifies how a Steenrod module should be constructed.

Constants

RED_ANSI_CODE

STATIC_MODULES_PATH

WHITE_ANSI_CODE

Traits

HashMapTuple

Functions

ascii_num

construct
This constructs a resolution resolving a module according to the specifications

iter_stems
Iterate through all pairs (s, f, t) such that f = t - s, s <= max_s and t <= max_t

iter_stems_f
Iterate through all pairs (s, f, t) such that f = t - s, s <= max_s and f <= max_f

load_module_json

parse_module_name

print_resolution_color