Function ext::utils::construct_s_2

pub fn construct_s_2<T: AsRef<Path>>(
    algebra: &str, 
    save_file: Option<T>
) -> Resolution<CCC>

A function that constructs the resolution of S_2 over the algebra. If save_file points to an existent file, then it loads the resolution from the save file. It is not an error to supply a non-existent file, but it is an error to supply an invalid save file.

This is used for various examples and tests as a shorthand.