ext::yoneda

Function try_yoneda_representative_element

pub fn try_yoneda_representative_element<CC>(
    cc: Arc<CC>,
    b: Bidegree,
    class: &[u32],
) -> Result<Yoneda<CC>>
where
    CC: FreeChainComplex + AugmentedChainComplex<ChainMap = FreeModuleHomomorphism<<<CC as AugmentedChainComplex>::TargetComplex as ChainComplex>::Module>>,
    CC::TargetComplex: BoundedChainComplex,
    CC::Algebra: GeneratedAlgebra,

Fallible variant of yoneda_representative_element: this holds the actual computation.

It validates the cheaply-checkable preconditions up front and returns an error instead of panicking when they are violated, namely:

the bidegree b must have been computed in cc, and

the length of class must match the number of generators of cc in bidegree b.

The internal Euler-characteristic / lift sanity checks (mathematical invariants that can only be verified after replaying the computation) are likewise surfaced as errors rather than panicking. yoneda_representative_element is try_yoneda_representative_element(..).unwrap().