Crate ext

ext-rs is a collection of libraries for doing homological algebra over $F_p$. The main and original purpose is to compute Ext of a Steenrod module, but the library is written to be sufficiently generic to deal with more general applications.

Subcrates

This contains a number of sub-crates, which each have their own documentation. A brief overview is as follows:

Name	Description

algebra	This defines algebras, modules and module homomorphisms

bivec	This is a small crate that provides BiVec - a variant of Vec indexed by an i32 whose starting index may be non-zero.

chart	This provides some APIs for generating charts

error	Our bespoke error library

fp	This implements linear algebra over $\mathbb{F}_p$, as well as general helper functions about primes.

once	This provides OnceVec and OnceBiVec, a push-only vector with non-blocking reads. This models some partially computed infinite data structure, and we think of pushing as simply finding out more of this infinite data structure instead of genuinely mutating it.

query	This contains some helper functions for a command line interface.

saveload	This provides an interface for saving and loading resolutions and other data.

thread_token	This provides some concurrency primitives

Examples

The library also comes with a number of “example” binaries that use the library for various purposes. These can be used directly to perform specific computations, or act as actual examples for how to use the library.

General usage guides can be found in the README; click on the links below for more individual details.

Name	Description

algebra_dim	Print the dimension of the Steenrod algebra in each degree.

bruner	Compare our basis with Bruner’s.

define_module	Interactively define a Steenrod module.

differentials	Print all differentials in the minimal resolution.

filtration_one	Print all filtration one products.

hidden	Compute hidden extensions using the output of various other examples.

lift_hom	Compute the map on Ext induced by a module homomorphism.

massey	Compute Massey products.

num_gens	Compute the dimension of Ext in each bidegree.

resolution_size	Compute the size of the minimal resolution in each bidegree

resolve	Resolve a module to a fixed (s, t) and potentially save the resolution.

resolve_through_stem	Resolve a module to a fixed (s, n) and potentially save the resolution.

save_bruner	Save the resolution in the format used by Bruner’s ext.

secondary	Compute $d_2$ differentials using the secondary Steenrod algebra.

steenrod	Compute Steenrod operations in Ext.

tensor	Compute the tensor product of two modules.

yoneda	Compute a Yoneda representative of an Ext class.

Modules

chain_complex

resolution

resolution_homomorphism

secondary

utils

yoneda

Type Definitions

CCC