ext
Module chain_complex

Re-exports

pub use chain_homotopy::ChainHomotopy;

Modules

chain_homotopy 🔒

finite_chain_complex 🔒

Structs

ChainMap

chain_maps is required to be non-empty

FiniteAugmentedChainComplex

FiniteChainComplex

StemIterator

An iterator returned by ChainComplex::iter_stem

Enums

ChainComplexGrading

Traits

AugmentedChainComplex

An augmented chain complex is a map of chain complexes C -> D that is a quasi-isomorphism. We usually think of C as a resolution of D. The chain map must be a map of degree shift 0.

BoundedChainComplex

A bounded chain complex is a chain complex C for which C_s = 0 for all s >= max_s

ChainComplex

A chain complex is defined to start in degree 0. The min_degree is the min_degree of the modules in the chain complex, all of which must be the same.

FreeChainComplex