Trait ext::chain_complex::AugmentedChainComplex

pub trait AugmentedChainComplex: ChainComplex {
    type TargetComplex: ChainComplex<Algebra = Self::Algebra>;
    type ChainMap: ModuleHomomorphism<Source = Self::Module, Target = <Self::TargetComplex as ChainComplex>::Module>;

    // Required methods
    fn target(&self) -> Arc<Self::TargetComplex>;
    fn chain_map(&self, s: u32) -> Arc<Self::ChainMap>;
}

An augmented chain complex is a map of chain complexes C -> D that is a quasi-isomorphism. We usually think of C as a resolution of D. The chain map must be a map of degree shift 0.

Required Associated Types

type TargetComplex: ChainComplex<Algebra = Self::Algebra>

type ChainMap: ModuleHomomorphism<Source = Self::Module, Target = <Self::TargetComplex as ChainComplex>::Module>

Required Methods

fn target(&self) -> Arc<Self::TargetComplex>

fn chain_map(&self, s: u32) -> Arc<Self::ChainMap>

Object Safety

This trait is not object safe.