Trait ext::chain_complex::ChainComplex

pub trait ChainComplex: Send + Sync {
    type Algebra: Algebra;
    type Module: Module<Algebra = Self::Algebra>;
    type Homomorphism: ModuleHomomorphism<Source = Self::Module, Target = Self::Module>;

    fn algebra(&self) -> Arc<Self::Algebra>;

    fn min_degree(&self) -> i32;

    fn zero_module(&self) -> Arc<Self::Module>;

    fn module(&self, homological_degree: u32) -> Arc<Self::Module>;

    fn differential(&self, s: u32) -> Arc<Self::Homomorphism>;

    fn has_computed_bidegree(&self, s: u32, t: i32) -> bool;

    fn compute_through_bidegree(&self, s: u32, t: i32);

    fn next_homological_degree(&self) -> u32;

    fn set_homology_basis(
        &self, 
        homological_degree: u32, 
        internal_degree: i32, 
        homology_basis: Vec<usize>
    );

    fn homology_basis(
        &self, 
        homological_degree: u32, 
        internal_degree: i32
    ) -> &Vec<usize>;

    fn max_homology_degree(&self, homological_degree: u32) -> i32;

    fn prime(&self) -> ValidPrime { ... }

    fn compute_through_bidegree_concurrent(
        &self, 
        s: u32, 
        t: i32, 
        bucket: &TokenBucket
    ) { ... }

    fn compute_homology_through_bidegree(
        &self, 
        homological_degree: u32, 
        internal_degree: i32
    ) { ... }

    fn homology_dimension(
        &self, 
        homological_degree: u32, 
        internal_degree: i32
    ) -> usize { ... }

    fn homology_gen_to_cocyle(
        &self, 
        result: &mut FpVector, 
        coeff: u32, 
        homological_degree: u32, 
        internal_degree: i32, 
        index: usize
    ) { ... }

    fn compute_homology(&self, homological_degree: u32, internal_degree: i32) { ... }

    fn iter_stem(&self) -> StemIterator<'_, Self> { ... }

}

A chain complex is defined to start in degree 0. The min_degree is the min_degree of the modules in the chain complex, all of which must be the same.

Associated Types

type Algebra: Algebra

type Module: Module<Algebra = Self::Algebra>

type Homomorphism: ModuleHomomorphism<Source = Self::Module, Target = Self::Module>

Loading content...
Required methods

fn algebra(&self) -> Arc<Self::Algebra>

fn min_degree(&self) -> i32

fn zero_module(&self) -> Arc<Self::Module>

fn module(&self, homological_degree: u32) -> Arc<Self::Module>

fn differential(&self, s: u32) -> Arc<Self::Homomorphism>

This returns the differential starting from the sth module.

fn has_computed_bidegree(&self, s: u32, t: i32) -> bool

If the complex has been computed at bidegree (s, t). This means the module has been computed at (s, t), and so has the differential at (s, t). In the case of a free module, the target of the differential, namely the bidegree (s - 1, t), need not be computed, as long as all the generators hit by the differential have already been computed.

fn compute_through_bidegree(&self, s: u32, t: i32)

Ensure all bidegrees less than or equal to (s, t) have been computed

fn next_homological_degree(&self) -> u32

The first s such that self.module(s) is not defined.

fn set_homology_basis(
&self,
homological_degree: u32,
internal_degree: i32,
homology_basis: Vec<usize>
)

fn homology_basis(
&self,
homological_degree: u32,
internal_degree: i32
) -> &Vec<usize>

fn max_homology_degree(&self, homological_degree: u32) -> i32

Loading content...
Provided methods

fn prime(&self) -> ValidPrime

fn compute_through_bidegree_concurrent(
&self,
s: u32,
t: i32,
bucket: &TokenBucket
)

A concurrent version of compute_through_bidegree_concurrent. This defaults to the non-concurrent version

fn compute_homology_through_bidegree(
&self,
homological_degree: u32,
internal_degree: i32
)

fn homology_dimension(
&self,
homological_degree: u32,
internal_degree: i32
) -> usize

fn homology_gen_to_cocyle(
&self,
result: &mut FpVector,
coeff: u32,
homological_degree: u32,
internal_degree: i32,
index: usize
)

fn compute_homology(&self, homological_degree: u32, internal_degree: i32)

fn iter_stem(&self) -> StemIterator<'_, Self>

Iterate through all defind bidegrees in increasing order of stem. The return values are of the form (s, n, t).

Loading content...