Trait ext::chain_complex::FreeChainComplex

pub trait FreeChainComplex<const U: bool = false>: ChainComplex<Module = MuFreeModule<U, <Self as ChainComplex>::Algebra>, Homomorphism = MuFreeModuleHomomorphism<U, MuFreeModule<U, <Self as ChainComplex>::Algebra>>>where
    <Self as ChainComplex>::Algebra: MuAlgebra<U>,{
    // Provided methods
    fn graded_dimension_string(&self) -> String { ... }
    fn to_sseq(&self) -> Sseq<Adams> { ... }
    fn filtration_one_products(&self, op_deg: i32, op_idx: usize) -> Product { ... }
    fn filtration_one_product(
        &self,
        op_deg: i32,
        op_idx: usize,
        source: Bidegree
    ) -> Option<Vec<Vec<u32>>> { ... }
    fn number_of_gens_in_bidegree(&self, b: Bidegree) -> usize { ... }
    fn cocycle_string(&self, gen: BidegreeGenerator, compact: bool) -> String { ... }
}

Provided Methods

fn graded_dimension_string(&self) -> String

fn to_sseq(&self) -> Sseq<Adams>

fn filtration_one_products(&self, op_deg: i32, op_idx: usize) -> Product

fn filtration_one_product( &self, op_deg: i32, op_idx: usize, source: Bidegree ) -> Option<Vec<Vec<u32>>>

Computes the filtration one product.

Returns

If the chain complex is stable, this always returns Some. If it is unstable, this returns None if the product is not defined.

fn number_of_gens_in_bidegree(&self, b: Bidegree) -> usize

fn cocycle_string(&self, gen: BidegreeGenerator, compact: bool) -> String