ext::chain_complex
Trait FreeChainComplex

pub trait FreeChainComplex<const U: bool = false>: ChainComplex<Module = MuFreeModule<U, <Self as ChainComplex>::Algebra>, Homomorphism = MuFreeModuleHomomorphism<U, MuFreeModule<U, <Self as ChainComplex>::Algebra>>>
where
    <Self as ChainComplex>::Algebra: MuAlgebra<U>,
{
    // Provided methods
    fn graded_dimension_string(&self) -> String { ... }
    fn to_sseq(&self) -> Sseq<Adams> { ... }
    fn filtration_one_products(&self, op_deg: i32, op_idx: usize) -> Product { ... }
    fn filtration_one_product(
        &self,
        op_deg: i32,
        op_idx: usize,
        source: Bidegree,
    ) -> Option<Vec<Vec<u32>>> { ... }
    fn number_of_gens_in_bidegree(&self, b: Bidegree) -> usize { ... }
    fn iter_nonzero_stem(&self) -> impl Iterator<Item = Bidegree> + '_ { ... }
    fn boundary_string(&self, gen: BidegreeGenerator, compact: bool) -> String { ... }
}

Provided Methods

fn graded_dimension_string(&self) -> String

fn to_sseq(&self) -> Sseq<Adams>

fn filtration_one_products(&self, op_deg: i32, op_idx: usize) -> Product

fn filtration_one_product( &self, op_deg: i32, op_idx: usize, source: Bidegree, ) -> Option<Vec<Vec<u32>>>

Computes the filtration one product.

Returns

If the chain complex is stable, this always returns Some. If it is unstable, this returns None if the product is not defined.

fn number_of_gens_in_bidegree(&self, b: Bidegree) -> usize

fn iter_nonzero_stem(&self) -> impl Iterator<Item = Bidegree> + '_

Iterate through all nonzero bidegrees in increasing order of stem.

fn boundary_string(&self, gen: BidegreeGenerator, compact: bool) -> String

Get a string representation of d(gen), where d is the differential of the resolution.

Object Safety

This trait is not object safe.