Module ext::nassau

This module implements Nassau’s algorithm. The main export is the Resolution object, which is a resolution of the sphere at the prime 2 using Nassau’s algorithm. It aims to provide an API similar to resolution::Resolution. From an API point of view, the main difference between the two is that our Resolution is a chain complex over MilnorAlgebra over SteenrodAlgebra.

To make use of this resolution in the example scripts, enable the nassau feature. This will cause utils::query_module to return the Resolution from this module instead of resolution. There is no formal polymorphism involved; the feature changes the return type of the function. While this is an incorrect use of features, we find that this the easiest way to make all scripts support both types of resolutions.

Structs

MilnorSubalgebra 🔒

A Milnor subalgebra to be used in Nassau’s algorithm. This is equipped with an ordering of the signature as in Lemma 2.4 of the paper.

Resolution

A resolution of S_2 using Nassau’s algorithm. This aims to have an API similar to that of resolution::Resolution. From an API point of view, the main difference between the two is that this is a chain complex over MilnorAlgebra over SteenrodAlgebra.

SenderData 🔒

See resolution::SenderData. This differs by not having the new field.

SignatureIterator 🔒

See MilnorSubalgebra::iter_signatures.

SubalgebraIterator 🔒

An iterator that iterates through a sequence of MilnorSubalgebra of increasing size. This is used by MilnorSubalgebra::optimal_for to find the largest subalgebra in this sequence that is applicable to a bidegree.

Enums

Magic 🔒

Some magic constants used in the save file

Constants

MAX_NEW_GENS 🔒