Struct ext::nassau::MilnorSubalgebra

pub struct MilnorSubalgebra {
    profile: Vec<u8>,
}

A Milnor subalgebra to be used in Nassau’s algorithm. This is equipped with an ordering of the signature as in Lemma 2.4 of the paper.

To simplify implementation, we pick the ordering so that the (reverse) lexicographic ordering in Lemma 2.4 is just the (reverse) lexicographic ordering of the P parts. This corresponds to the ordering of $\mathcal{P}$ where $P^s_t < P^{s’}_t$ if $s < s’$).

Fields
profile: Vec<u8>
Implementations

impl MilnorSubalgebra

pub const INFINITY: u8

This should be used when you want an entry of the profile to be infinity

pub fn new(profile: Vec<u8>) -> Self

pub fn zero_algebra() -> Self

pub fn has_signature(
&self,
elt: &MilnorBasisElement,
signature: &[PPartEntry]
) -> bool

Computes the signature of an element

pub fn zero_signature(&self) -> Vec<PPartEntry>

pub fn signature_mask<'a>(
&'a self,
module: &'a FreeModule<MilnorAlgebra>,
degree: i32,
signature: &'a [PPartEntry]
) -> impl Iterator<Item = usize> + 'a

Give a list of basis elements in degree degree that has signature signature.

pub fn signature_matrix(
&self,
hom: &FreeModuleHomomorphism<FreeModule<MilnorAlgebra>>,
degree: i32,
signature: &[PPartEntry]
) -> Matrix

Get the matrix of a free module homomorphism when restricted to the subquotient given by the signature.

pub fn iter_signatures(
&self,
degree: i32
) -> impl Iterator<Item = Vec<PPartEntry>> + '_

Iterate through all signatures of this algebra that contain elements of degree at most degree (inclusive). This skips the initial zero signature

fn top_degree(&self) -> i32

fn optimal_for(s: u32, t: i32) -> MilnorSubalgebra