Struct ext::nassau::MilnorSubalgebra

struct MilnorSubalgebra {
    profile: Vec<u8>,
}

A Milnor subalgebra to be used in Nassau’s algorithm. This is equipped with an ordering of the signature as in Lemma 2.4 of the paper.

To simplify implementation, we pick the ordering so that the (reverse) lexicographic ordering in Lemma 2.4 is just the (reverse) lexicographic ordering of the P parts. This corresponds to the ordering of $\mathcal{P}$ where $P^s_t < P^{s’}_t$ if $s < s’$).

Fields
profile: Vec<u8>
Implementations

impl MilnorSubalgebra

const INFINITY: u8

This should be used when you want an entry of the profile to be infinity

fn new(profile: Vec<u8>) -> Self

fn zero_algebra() -> Self

The algebra with trivial profile, corresponding to the trivial algebra.

fn has_signature(&self, ppart: &[PPartEntry], signature: &[PPartEntry]) -> bool

Computes the signature of an element

fn zero_signature(&self) -> Vec<PPartEntry>

fn signature_mask<'a>(
&'a self,
algebra: &'a MilnorAlgebra,
module: &'a FreeModule<MilnorAlgebra>,
degree: i32,
signature: &'a [PPartEntry]
) -> impl Iterator<Item = usize> + 'a

Give a list of basis elements in degree degree that has signature signature.

This requires passing the algebra for borrow checker reasons.

fn signature_matrix(
&self,
hom: &FreeModuleHomomorphism<FreeModule<MilnorAlgebra>>,
degree: i32,
signature: &[PPartEntry]
) -> Matrix

Get the matrix of a free module homomorphism when restricted to the subquotient given by the signature.

fn iter_signatures(
&self,
degree: i32
) -> impl Iterator<Item = Vec<PPartEntry>> + '_

Iterate through all signatures of this algebra that contain elements of degree at most degree (inclusive). This skips the initial zero signature.

fn top_degree(&self) -> i32

fn optimal_for(s: u32, t: i32) -> MilnorSubalgebra

fn to_bytes(&self, buffer: &mut impl Write) -> Result<()>

fn from_bytes(data: &mut impl Read) -> Result<Self>

fn signature_to_bytes(
signature: &[PPartEntry],
buffer: &mut impl Write
) -> Result<()>

fn signature_from_bytes(&self, data: &mut impl Read) -> Result<Vec<PPartEntry>>