Struct ext::nassau::Resolution

pub struct Resolution {
    lock: Mutex<()>,
    modules: OnceVec<Arc<FreeModule<MilnorAlgebra>>>,
    zero_module: Arc<FreeModule<MilnorAlgebra>>,
    differentials: OnceVec<Arc<FreeModuleHomomorphism<FreeModule<MilnorAlgebra>>>>,
}

A resolution of a chain complex.

Fields
lock: Mutex<()>modules: OnceVec<Arc<FreeModule<MilnorAlgebra>>>zero_module: Arc<FreeModule<MilnorAlgebra>>differentials: OnceVec<Arc<FreeModuleHomomorphism<FreeModule<MilnorAlgebra>>>>
Implementations

impl Resolution

pub const fn prime(&self) -> ValidPrime

pub fn new() -> Self

fn extend_through_degree(&self, max_s: u32)

This function prepares the Resolution object to perform computations up to the specified s degree. It does not perform any computations by itself. It simply lengthens the OnceVecs modules, chain_maps, etc. to the right length.

fn step_resolution_with_subalgebra(
&self,
s: u32,
t: i32,
subalgebra: MilnorSubalgebra
)

fn step_resolution(&self, s: u32, t: i32)

pub fn compute_through_stem(&self, max_s: u32, max_n: i32)

This function resolves up till a fixed stem instead of a fixed t.