Struct ext::resolution::Resolution

pub struct Resolution<CC: ChainComplex> {
    lock: Mutex<()>,
    complex: Arc<CC>,
    modules: OnceVec<Arc<FreeModule<<CC::Module as Module>::Algebra>>>,
    zero_module: Arc<FreeModule<<CC::Module as Module>::Algebra>>,
    chain_maps: OnceVec<Arc<FreeModuleHomomorphism<CC::Module>>>,
    differentials: OnceVec<Arc<FreeModuleHomomorphism<FreeModule<<CC::Module as Module>::Algebra>>>>,
    kernels: OnceBiVec<Mutex<Option<Subspace>>>,
}

A resolution of a chain complex.

Fields
lock: Mutex<()>complex: Arc<CC>modules: OnceVec<Arc<FreeModule<<CC::Module as Module>::Algebra>>>zero_module: Arc<FreeModule<<CC::Module as Module>::Algebra>>chain_maps: OnceVec<Arc<FreeModuleHomomorphism<CC::Module>>>differentials: OnceVec<Arc<FreeModuleHomomorphism<FreeModule<<CC::Module as Module>::Algebra>>>>kernels: OnceBiVec<Mutex<Option<Subspace>>>

For each internal degree, store the kernel of the most recently calculated chain map as returned by generate_old_kernel_and_compute_new_kernel, to be used if we run compute_through_degree again.

Implementations

impl<CC: ChainComplex> Resolution<CC>

pub fn new(complex: Arc<CC>) -> Self

pub fn extended_degree(&self) -> (u32, i32)

fn extend_through_degree(&self, max_s: u32, max_t: i32)

This function prepares the Resolution object to perform computations up to the specified s degree. It does not perform any computations by itself. It simply lengthens the OnceVecs modules, chain_maps, etc. to the right length.

fn step_resolution(&self, s: u32, t: i32)

Call our resolution $X$, and the chain complex to resolve $C$. This is a legitimate resolution if the map $f: X \to C$ induces an isomorphism on homology. This is the same as saying the cofiber is exact. The cofiber is given by the complex

$$ X_s \oplus C_{s+1} \to X_{s-1} \oplus C_s \to X_{s-2} \oplus C_{s-1} \to \cdots $$

where the differentials are given by

$$ \begin{pmatrix} d_X & 0 \\ (-1)^s f & d_C \end{pmatrix} $$

Our method of producing $X_{s, t}$ and the chain maps are as follows. Suppose we have already built the chain map and differential for $X_{s-1, t}$ and $X_{s, t-1}$. Since $X_s$ is a free module, the generators in degree $< t$ gives us a bunch of elements in $X_s$ already, and we know exactly where they get mapped to. Let $T$ be the $\mathbb{F}_p$ vector space generated by these elements. Then we already have a map

$$ T \to X_{s-1, t} \oplus C_{s, t}$$

and we know this hits the kernel of the map

$$ D = X_{s-1, t} \oplus C_{s, t} \to X_{s-2, t} \oplus C_{s-1, t}. $$

What we need to do now is to add generators to $X_{s, t}$ to hit the entirity of this kernel. Note that we don’t have to do this. Some of the elements in the kernel might be hit by $C_{s+1, t}$ and we don’t have to hit them, but we opt to add generators to hit it anyway.

If we do it this way, then we know the composite of the map

$$ T \to X_{s-1, t} \oplus C_{s, t} \to C_{s, t} $$

has to be surjective, since the image of $C_{s, t}$ under $D$ is also in the image of $X_{s-1, t}$. So our first step is to add generators to $X_{s, t}$ such that this composite is surjective.

After adding these generators, we need to decide where to send them to. We know their values in the $C_{s, t}$ component, but we need to use a quasi-inverse to find the element in $X_{s-1, t}$ that hits the corresponding image of $C_{s, t}$. This tells us the $X_{s-1, t}$ component.

Finally, we need to add further generators to $X_{s, t}$ to hit all the elements in the kernel of

$$ X_{s-1, t} \to X_{s-2, t} \oplus C_{s-1, t}. $$

This kernel was recorded by the previous iteration of the method in old_kernel, so this step is doable as well.

Note that if we add our new generators conservatively, then the kernel of the maps

$$ \begin{aligned} T &\to X_{s-1, t} \oplus C_{s, t} \\ X_{s, t} &\to X_{s-1, t} \oplus C_{s, t} \end{aligned} $$ agree.

In the code, we first row reduce the matrix of the map from $T$. This lets us record the kernel which is what the function returns at the end. This computation helps us perform the future steps since we need to know about the cokernel of this map.

Arguments

s - The s degree to calculate

t - The t degree to calculate

fn step_resolution_phony(&self, s: u32, t: i32)

pub fn cocycle_string(&self, hom_deg: u32, int_deg: i32, idx: usize) -> String

pub fn complex(&self) -> Arc<CC>

pub fn number_of_gens_in_bidegree(&self, s: u32, t: i32) -> usize

pub fn prime(&self) -> ValidPrime

pub fn compute_through_bidegree_concurrent_with_callback(
&self,
max_s: u32,
max_t: i32,
bucket: &TokenBucket,
cb: impl FnMut(u32, i32)
)

pub fn compute_through_bidegree_with_callback(
&self,
max_s: u32,
max_t: i32,
cb: impl FnMut(u32, i32)
)

pub fn compute_through_stem(&self, max_s: u32, max_n: i32)

This function resolves up till a fixed stem instead of a fixed t. It is an error to attempt to resolve further after this is called, and will result in a deadlock.

pub fn filtration_one_product(
&self,
op_deg: i32,
op_idx: usize,
target_s: u32,
target_t: i32
) -> Option<Vec<Vec<u32>>>

Computes the filtration one product. This returns None if the source or target is out of range.

pub fn compute_through_stem_concurrent(
&self,
max_s: u32,
max_n: i32,
bucket: &TokenBucket
)

A concurrent version of Resolution::compute_through_stem