Struct ext::secondary::SecondaryHomotopy

pub struct SecondaryHomotopy<A: PairAlgebra> {
    pub source: Arc<FreeModule<A>>,
    pub target: Arc<FreeModule<A>>,
    pub shift_t: i32,
    pub(crate) composites: OnceBiVec<Vec<SecondaryComposite<A>>>,
    pub homotopies: FreeModuleHomomorphism<FreeModule<A>>,
    hit_generator: bool,
}

Fields
source: Arc<FreeModule<A>>target: Arc<FreeModule<A>>shift_t: i32

output_t = input_t - shift_t

composites: OnceBiVec<Vec<SecondaryComposite<A>>>

gen_deg -> gen_idx -> composite

homotopies: FreeModuleHomomorphism<FreeModule<A>>

gen_deg -> gen_idx -> homotopy

hit_generator: bool
Implementations

impl<A: PairAlgebra + Send + Sync> SecondaryHomotopy<A>

pub fn new(
source: Arc<FreeModule<A>>,
target: Arc<FreeModule<A>>,
shift_t: i32,
hit_generator: bool
) -> Self

pub fn add_composite(
&self,
s: u32,
degree: i32,
maps: CompositeData<A>,
dir: Option<&Path>
)

Add composites up to and including the specified degree

pub fn act(
&self,
result: SliceMut<'_>,
coeff: u32,
elt_degree: i32,
elt: Slice<'_>,
full: bool
)

Compute the image of an element in the source under the homotopy, writing the result in result. It is assumed that the coefficients of generators are zero in op.

Arguments

full: Whether to include the action of the homotopy part as well

pub fn composite(&self, gen_deg: i32, gen_idx: usize) -> &SecondaryComposite<A>