Trait ext::secondary::SecondaryLift

pub trait SecondaryLift: Sync {
    type Algebra: PairAlgebra;
    type Source: FreeChainComplex<Algebra = Self::Algebra>;
    type Target: FreeChainComplex<Algebra = Self::Algebra>;
    type Underlying;

    const HIT_GENERATOR: bool = false;

    fn underlying(&self) -> Arc<Self::Underlying>;
    fn algebra(&self) -> Arc<Self::Algebra>;
    fn source(&self) -> Arc<Self::Source>;
    fn target(&self) -> Arc<Self::Target>;
    fn shift_t(&self) -> i32;
    fn shift_s(&self) -> u32;
    fn max_s(&self) -> u32;
    fn max_t(&self, s: u32) -> i32;
    fn homotopies(&self) -> &OnceBiVec<SecondaryHomotopy<Self::Algebra>>;
    fn intermediates(&self) -> &DashMap<(u32, i32, usize), FpVector>;
    fn save_dir(&self) -> Option<&Path>;
    fn compute_intermediate(&self, s: u32, t: i32, idx: usize) -> FpVector;
    fn composite(&self, s: u32) -> CompositeData<Self::Algebra>;

    fn prime(&self) -> ValidPrime { ... }
    fn initialize_homotopies(&self) { ... }
    fn compute_composites(&self) { ... }
    fn get_intermediate(&self, s: u32, t: i32, idx: usize) -> FpVector { ... }
    fn compute_partial(&self, s: u32) { ... }
    fn compute_intermediates(&self) { ... }
    fn compute_homotopy_step(&self, s: u32, t: i32) -> Range<i32> { ... }
    fn compute_homotopies(&self) { ... }
    fn extend_all(&self) { ... }

}

When lifting a thing to its secondary version, often what we have to do is to specify an explicit homotopy to witnesses that some equation holds. For example, to lift a chain complex, we need a homotopy witnessing the fact that $d^2 \simeq 0$. This homotopy in turn is required to satisfy certain recursive relations.

To specify this lifting problem, one needs to supply two pieces of data. First is the equation that we are trying to witness, which is usually of the form

$$ \sum_i c_i f_i g_i = 0, $$

where $f_i$ and $g_i$ are free module homomorphisms and $c_i$ are constants. This is specified by SecondaryLift::composite.

The next is a compatibility equation, which restricts the τ part of the null-homotopy, and is usually of the form

$$ dh = hd + \mathrm{stuff} $$

The τ part of $hd + \mathrm{stuff}$ is known as the intermediate data, and is what SecondaryLift::compute_intermediate returns.

Required Associated Types

type Algebra: PairAlgebra

type Source: FreeChainComplex<Algebra = Self::Algebra>

type Target: FreeChainComplex<Algebra = Self::Algebra>

type Underlying

Provided Associated Constants

const HIT_GENERATOR: bool = false

Whether the composite can hit generators. This is true for SecondaryChainHomotopy and false for the rest. This is important because for SecondaryResolution, we don’t actually know all the generators if we resolve up to a stem. So in composites for SecondaryResolution, we need to ignore target generators of the same degree uniformly.

Required Methods

fn underlying(&self) -> Arc<Self::Underlying>

fn algebra(&self) -> Arc<Self::Algebra>

fn source(&self) -> Arc<Self::Source>

fn target(&self) -> Arc<Self::Target>

fn shift_t(&self) -> i32

fn shift_s(&self) -> u32

fn max_s(&self) -> u32

Exclusive max s

fn max_t(&self, s: u32) -> i32

Exclusive max t

fn homotopies(&self) -> &OnceBiVec<SecondaryHomotopy<Self::Algebra>>

fn intermediates(&self) -> &DashMap<(u32, i32, usize), FpVector>

fn save_dir(&self) -> Option<&Path>

fn compute_intermediate(&self, s: u32, t: i32, idx: usize) -> FpVector

fn composite(&self, s: u32) -> CompositeData<Self::Algebra>

Provided Methods

fn prime(&self) -> ValidPrime

fn initialize_homotopies(&self)

fn compute_composites(&self)

fn get_intermediate(&self, s: u32, t: i32, idx: usize) -> FpVector

fn compute_partial(&self, s: u32)

fn compute_intermediates(&self)

fn compute_homotopy_step(&self, s: u32, t: i32) -> Range<i32>

fn compute_homotopies(&self)

fn extend_all(&self)