Module ext::utils

Structs

Config

A config object is an object that specifies how a Steenrod module should be constructed.

QueryModuleResult

Enums

LoadQuasiInverseOption

Constants

RED_ANSI_CODE

STATIC_MODULES_PATH

WHITE_ANSI_CODE

Traits

HashMapTuple

Functions

ascii_num

construct

This constructs a resolution resolving a module according to the specifications

load_module_json

parse_module_name

print_element

Prints an element in the bidegree (n, s) to stdout. For example, [0, 2, 1] will be printed as 2 x_(n, s, 1) + x_(f, s, 2).

print_resolution_color

query_bucket

query_module

query_module_only

query_num_threads