ext

Module utils

A module containing various utility functions related to user interaction in some way.

Modules

logging 🔒

Structs

Config

A config object is an object that specifies how a Steenrod module should be constructed.

LogWriter

Constants

STATIC_MODULES_PATH 🔒

Functions

construct

This constructs a resolution resolving a module according to the specifications

construct_nassau

See construct

construct_standard

See construct

ext_tracing_subscriber

get_unit

Given a resolution, return a resolution of the unit.

init_logging

load_module_json

Load a module specification from a JSON file.

parse_module_name

Given a module specification string, load a json description of the module as described here.

query_module

Query the user for a module and a bidegree, and return a resolution resolved up to said bidegree.

query_module_only

Query the user for a module and its save directory. See here for details on the propmt format.

query_unstable_module

query_unstable_module_only

secondary_job

The value of the SECONDARY_JOB environment variable.

unicode_num

Given an n: usize, return a UTF-8 character that best depicts this number. If n < 9, then this is a UTF-8 when n many dots. If n = 9, then this is the number 9. Otherwise, it is *.

Type Aliases

QueryModuleResolution

The type returned by query_module. The value of this type depends on whether nassau is enabled. In any case, it is an augmented free chain complex over either SteenrodAlgebra or MilnorAlgebra and supports the compute_through_stem function.