Function ext::utils::iter_s_t

pub fn iter_s_t(
    f: &impl Fn(u32, i32) -> Range<i32> + Sync, 
    min_s: u32, 
    min_t: i32, 
    max_s: u32, 
    max_t: &impl Fn(u32) -> i32 + Sync
)

Given a function f(s, t), compute it for every s in [min_s, max_s] and every t in [min_t, max_t(s)]. Further, we only compute f(s, t) when f(s - 1, t') has been computed for all t' < t.

The function f should return a range starting from t and ending at the largest T such that f(s, t') has already been computed for every t' < T.

While iter_s_t could have had kept track of that data, it is usually the case that f would compute something and write it to a OnceBiVec, and OnceBiVec::push_ooo would return this range for us.

This uses rayon under the hood, and f should feel free to use further rayon parallelism.

Arguments:

max_s: This is exclusive

max_t: This is exclusive