Function ext::utils::query_module

pub fn query_module(
    algebra: Option<AlgebraType>, 
    load_quasi_inverse: bool
) -> Result<QueryModuleResolution>

Query the user for a module and a bidegree, and return a resolution resolved up to said bidegree. This is mainly a wrapper around query_module_only that also asks for the bidegree to resolve up to as well. The prompt of query_module_only is always set to "Module" when invoked through this function.