Type Alias ext::utils::QueryModuleResolution

pub type QueryModuleResolution = Resolution<FDModule<MilnorAlgebra>>;

The type returned by query_module. The value of this type depends on whether nassau is enabled. In any case, it is an augmented free chain complex over either SteenrodAlgebra or MilnorAlgebra and supports the compute_through_stem function.

Aliased Type

struct QueryModuleResolution {
    lock: Mutex<()>,
    name: String,
    max_degree: i32,
    modules: OnceVec<Arc<MuFreeModule<false, MilnorAlgebra>>>,
    zero_module: Arc<MuFreeModule<false, MilnorAlgebra>>,
    differentials: OnceVec<Arc<MuFreeModuleHomomorphism<false, MuFreeModule<false, MilnorAlgebra>>>>,
    target: Arc<FiniteChainComplex<FiniteDimensionalModule<MilnorAlgebra>>>,
    chain_maps: OnceVec<Arc<MuFreeModuleHomomorphism<false, FiniteDimensionalModule<MilnorAlgebra>>>>,
    save_dir: Option<PathBuf>,
}

Fields
lock: Mutex<()>name: Stringmax_degree: i32modules: OnceVec<Arc<MuFreeModule<false, MilnorAlgebra>>>zero_module: Arc<MuFreeModule<false, MilnorAlgebra>>differentials: OnceVec<Arc<MuFreeModuleHomomorphism<false, MuFreeModule<false, MilnorAlgebra>>>>target: Arc<FiniteChainComplex<FiniteDimensionalModule<MilnorAlgebra>>>chain_maps: OnceVec<Arc<MuFreeModuleHomomorphism<false, FiniteDimensionalModule<MilnorAlgebra>>>>save_dir: Option<PathBuf>