Function ext::yoneda::compute_kernel_image

fn compute_kernel_image<M: Module, F: ModuleHomomorphism, G: ModuleHomomorphism>(
    source: &QuotientModule<M>,
    augmentation_map: Option<&F>,
    preserve_map: Option<&G>,
    keep: Option<&Subspace>,
    t: i32
) -> (Matrix, Matrix)
where
    M::Algebra: GeneratedAlgebra,

This function does the following computation:

Given the source module source and a subspace keep, the function returns the subspace of all elements in source of degree t that are killed by all non-trivial actions of the algebra, followed by a list of elements that span the intersection between this subspace and keep.

If keep is None, it is interpreted as the empty subspace.